Update Rust edition, deps, fix warns/lints before reviving the project
Considering using the project to try a few things.. modernize it
slightly before making changes.

It doesn't run anymore (assertion error in a cranelift dependency) but
that's fine for what I'm planning to do.

+++ src/utils.rs
@@ -64,13 +64,13 @@ pub fn base62_encode(uniq: Uniq, w: &mut dyn Write) -> fmt::Result {
     let mut c = uniq.0.get() as usize;
     loop {
         if c < 62 {
-            w.write_char(char::from(BASE62_CHARS[c as usize]))?;
+            w.write_char(char::from(BASE62_CHARS[c]))?;
             break;
         }
 
         let q = c % 62;
         let r = c / 62;
-        w.write_char(char::from(BASE62_CHARS[q as usize]))?;
+        w.write_char(char::from(BASE62_CHARS[q]))?;
         c = r;
     }
